CUDA decompression first version
This version allows CUDA variable rate decompression. In order to use this feature, the input file must have an offset header with the chunk offsets on top of the data. This header can currently only be encoded with the OpenMP variable rate patches. These patches will be combined soon. Currently there are no plans to work on CUDA variable rate compression.