OpenMP and CUDA variable rate decompression
This patch adds support for OpenMP fixed rate, accuracy and precision decompression, and CUDA fixed accuracy and precision decompression. It does that by generating an "offset table" during the compression phase, and then using that table for performing parallel variable-rate decoding. The patch breaks the following tests:

testOmp: The test is broken because it expects a zero output when attempting OpenMP decompression, as this was previously not implemented. This part of the test should be removed.

The patch has been tested locally for OpenMP 1D, 2D, 3D (4D is not tested yet) and CUDA 1D, 2D and 3D.
The tests will be added in a follow up once we align on the implementation.